fix Ci
